替检接口开发

+++ fxkc-modules/fxkc-peis/src/main/java/org/fxkc/peis/domain/TjRegister.java
@@ -407,7 +407,7 @@ public class TjRegister extends TenantEntity {
     /**
      * 被替检人年龄
      */
-    private Long replaceAge;
+    private Integer replaceAge;
 
     /**
      * 排班id
