+++ fxkc-modules/fxkc-peis/src/main/java/org/fxkc/peis/domain/bo/TjRegisterBo.java
@@ -435,7 +435,7 @@ public class TjRegisterBo extends BaseEntity {
     /**
      * 被替检人年龄
      */
-    private Long replaceAge;
+    private Integer replaceAge;
 
     /**
      * 排班id
